refactor(client): step 1 -- client.py -> client/ package, move module-level helpers to _transfer.py (#520)
Pure movement, zero behavior change. Converts the 3,962-LOC single-file
client.py into a client/ package. The ~525 LOC of module-level helpers
(cloud up/download plumbing, query-error parsers, URL-safety, inline
Query Service pagination, Queue poll scheduling) move verbatim into
client/_transfer.py. The whole KeboolaClient class stays intact in
client/_client.py for now (peeled into mixins in later steps).

client/__init__.py re-exports the 9 public names so
`from keboola_agent_cli.client import X` is unchanged for all importers,
plus `time`, `QUERY_RESULTS_PAGE_SIZE` and `_CloudDownloader` which
existing tests reach as `keboola_agent_cli.client.<name>` (patch targets
and a page-size assertion).

Relative parent-package imports shift one level (`.constants` ->
`..constants`, etc.). A TYPE_CHECKING import of KeboolaClient in
_transfer.py keeps the `_collect_inline_results` forward-ref resolvable
for ty.

Test layout couplings updated (necessary consequence of a module split,
not a behavior change): the two workspace pagination tests and the page-
boundary test monkeypatch QUERY_RESULTS_PAGE_SIZE at its new reader
module (client._transfer); test_lib relaxes the InlineQueryResult
`__module__` equality to a client-package prefix check.

make check green: 4648 passed, 8 skipped. ty clean (3 pre-existing
unrelated warnings only).
